Se corrigen los seeder y el modelo Mascota para manejar correctamente la fecha de nacimiento y la edad de las mascotas.

+++ laravel12-react/app/Models/Mascota.php
@@ -61,7 +61,7 @@ class Mascota extends Model
     {
         static::saving(function ($mascota) {
             if ($mascota->fecha_nacimiento) {
-                $mascota->edad = Carbon::parse($mascota->fecha_nacimiento)->diffInYears(Carbon::now());
+                $mascota->edad = (int) Carbon::parse($mascota->fecha_nacimiento)->diffInYears(Carbon::now());
             }
         });
     }
